# ExpressFunctionApp/docs/swaggerv2.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Az Function Mock Web Service V2",
    "description": "Azure Function Mock Web Service V2"
  },
...
  "host": "azamwebsvc.azurewebsites.net",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/echo/{status}": {
      "post": {
        "summary": "Echoes back request",
        "description": "Echoes headers, querystring, and request body",
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "description": "Status code to be returned",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "echoRequest",
            "description": "Request body to be echoed",
            "schema": {
              "$ref": "#/definitions/echoRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/echoResponse"
            }
          }
        }
      }
    },
    "/echo-from-text/{status}": {
      "post": {
        "summary": "Echoes text as json",
        "description": "Echoes headers, querystring, and request body",
        "consumes": [
          "text/plain"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "description": "Status code to be returned",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "echoRequest",
            "x-ntx-summary": "Text request body to be echoed as json",
            "description": "{\"abc\": 123, \"def\": true, \"ghi\": \"value1\"}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/echoResponse"
            }
          }
        }
      }
    },
    "/files/errors/{status}": {
      "get": {
        "summary": "Download error",
        "description": "Returns error code instead of image for download",
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "description": "Error status code to be returned",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "images/jpeg"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "file": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/all-types": {
      "post": {
        "summary": "POST all variable types",
        "description": "POST all variable types",
        "parameters": [
          {
            "in": "body",
            "name": "rootBody",
            "required": true,
            "schema": {
              "$ref": "#/definitions/rootBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/allTypesOutputs"
            }
          }
        }
      }
    },
    "/all-types/object": {
      "get": {
        "summary": "Returns hardcoded object",
        "description": "Returns hardcoded object and text",
        "parameters": [
          {
            "in": "query",
            "name": "expected",
            "description": "Expected data to be returned",
            "required": false,
            "type": "string",
            "default": "body",
            "enum": [
              "body",
              "empty",
              "plaintext"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/allTypesOutputsObject"
            }
          }
        }
      }
    },
    "/all-types/array": {
      "get": {
        "summary": "Returns hardcoded array",
        "description": "Returns hardcoded array and text",
        "parameters": [
          {
            "in": "query",
            "name": "expected",
            "description": "Expected data to be returned",
            "required": false,
            "type": "string",
            "default": "body",
            "enum": [
              "body",
              "empty",
              "plaintext"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/allTypesOutputsArray"
            }
          }
        }
      }
    },
    "/all-types-stringified": {
      "post": {
        "summary": "POST all variable types and return stringified data",
        "description": "POST all variable types and return stringified data",
        "parameters": [
          {
            "in": "body",
            "name": "rootBody",
            "required": true,
            "schema": {
              "$ref": "#/definitions/rootBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "allTypesOutputsStringified": {
                  "type": "string",
                  "description": "Logged Request Inputs as String",
                  "x-ntx-summary": "Logged Request Inputs as String"
                }
              }
            }
          }
        }
      }
    },
    "/all-types-odata": {
      "post": {
        "summary": "POST all variable types as oData and return stringified data",
        "description": "POST all variable types as oData and return stringified data",
        "parameters": [
          {
            "in": "query",
            "name": "odata",
            "required": true,
            "type": "string",
            "x-ntx-query-builder": {
              "schema": {
                "type": "object",
                "properties": {
                  "textInput": {
                    "type": "string",
                    "description": "Text Input",
                    "x-ntx-summary": "Text Input"
                  },
                  "decimalInput": {
                    "type": "number",
                    "description": "Decimal Input",
                    "x-ntx-summary": "Decimal Input"
                  },
                  "integerInput": {
                    "type": "integer",
                    "description": "Integer Input",
                    "x-ntx-summary": "Integer Input"
                  },
                  "booleanInput": {
                    "type": "boolean",
                    "description": "Boolean Input",
                    "x-ntx-summary": "Boolean Input"
                  },
                  "datetimeInput": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Datetime Input",
                    "x-ntx-summary": "Datetime Input"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "allTypesOutputsStringified": {
                  "type": "string",
                  "description": "Logged Request Inputs as String",
                  "x-ntx-summary": "Logged Request Inputs as String"
                }
              }
            }
          }
        }
      }
    },
    "/all-types-nullable": {
      "get": {
        "summary": "GET all variable types with null value",
        "description": "GET all variable types with null value",
        "parameters": [
          {
            "in": "query",
            "name": "expected",
            "description": "Expected data to be returned",
            "required": false,
            "type": "string",
            "default": "validValues",
            "enum": [
              "validValues",
              "explicitNull",
              "missingProperty"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/allTypesNullableOutputsObject"
            }
          }
        }
      }
    },
    "/all-parameter-types/{string-path}/{integer-path}/{boolean-path}": {
      "post": {
        "summary": "POST all parameter types",
        "description": "POST all parameter types",
        "parameters": [
          {
            "in": "header",
            "name": "string_header",
            "x-ntx-summary": "Text Header Input",
            "type": "string"
          },
          {
            "in": "header",
            "name": "integer_header",
            "x-ntx-summary": "Integer Header Input",
            "type": "integer"
          },
          {
            "in": "header",
            "name": "boolean_header",
            "x-ntx-summary": "Boolean Header Input",
            "type": "boolean"
          },
          {
            "in": "path",
            "name": "string-path",
            "x-ntx-summary": "Text Path Input",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "integer-path",
            "x-ntx-summary": "Integer Path Input",
            "type": "integer",
            "required": true
          },
          {
            "in": "path",
            "name": "boolean-path",
            "x-ntx-summary": "Boolean Path Input",
            "type": "boolean",
            "required": true
          },
          {
            "in": "query",
            "name": "string_query",
            "description": "Text Query Input",
            "x-ntx-summary": "Text Query Input",
            "type": "string"
          },
          {
            "in": "query",
            "name": "integer_query",
            "description": "Integer Query Input",
            "x-ntx-summary": "Integer Query Input",
            "type": "integer"
          },
          {
            "in": "query",
            "name": "boolean_query",
            "description": "Boolean Query Input",
            "x-ntx-summary": "boolean Query Input",
            "type": "boolean"
          },
          {
            "in": "body",
            "name": "string_body",
            "description": "Request Body",
            "x-ntx-summary": "Request Body",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "description": "Text Property Body Input",
                  "x-ntx-summary": "Text Property Body Input"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/allParameterTypesRootOutput"
            }
          }
        }
      }
    },
    "/path-encoding/{string-path}": {
      "post": {
        "summary": "POST text route path encoding",
        "description": "Returns literal route path for testing encoding performed by the caller",
        "parameters": [
          {
            "in": "path",
            "name": "string-path",
            "x-ntx-summary": "Text Path Input",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "inputs": {
                  "type": "object",
                  "description": "Logged Request Inputs",
                  "x-ntx-summary": "Logged Request Inputs"
                },
                "path": {
                  "type": "string",
                  "description": "Text Path Output",
                  "x-ntx-summary": "Text Path Output"
                }
              }
            }
          }
        }
      }
    },
    "/query-encoding": {
      "post": {
        "summary": "POST text query string encoding",
        "description": "Returns literal querystring for testing encoding performed by the caller",
        "parameters": [
          {
            "in": "query",
            "name": "string_query",
            "description": "Text Query Input",
            "x-ntx-summary": "Text Query Input",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "inputs": {
                  "type": "object",
                  "description": "Logged Request Inputs",
                  "x-ntx-summary": "Logged Request Inputs"
                },
                "query": {
                  "type": "string",
                  "description": "Text Query Output",
                  "x-ntx-summary": "Text Query Output"
                }
              }
            }
          }
        }
      }
    },
    "/form-urlencoded/{string-path}/parsed": {
      "post": {
        "summary": "POST x-www-form-urlencoded content-type",
        "description": "Returns parsed form-urlencoded data",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "string-path",
            "x-ntx-summary": "Text Path Input",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "string",
            "x-ntx-summary": "Text Form Input",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "decimal",
            "x-ntx-summary": "Decimal Form Input",
            "type": "number"
          },
          {
            "in": "formData",
            "name": "integer",
            "x-ntx-summary": "Integer Form Input",
            "type": "integer"
          },
          {
            "in": "formData",
            "name": "boolean",
            "x-ntx-summary": "Boolean Form Input",
            "type": "boolean"
          },
          {
            "in": "formData",
            "name": "datetime",
            "x-ntx-summary": "Datetime Form Input",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "inputs": {
                  "type": "object",
                  "description": "Logged Request Inputs",
                  "x-ntx-summary": "Logged Request Inputs"
                },
                "outputs": {
                  "type": "object",
                  "properties": {
                    "textPathOutput": {
                      "type": "string",
                      "description": "Text Path Output",
                      "x-ntx-summary": "Text Path Output"
                    },
                    "textOutput": {
                      "type": "string",
                      "description": "Text Output",
                      "x-ntx-summary": "Text Output"
                    },
                    "decimalOutput": {
                      "type": "number",
                      "description": "Decimal Output",
                      "x-ntx-summary": "Decimal Output"
                    },
                    "integerOutput": {
                      "type": "integer",
                      "description": "Integer Output",
                      "x-ntx-summary": "Integer Output"
                    },
                    "booleanOutput": {
                      "type": "boolean",
                      "description": "Boolean Output",
                      "x-ntx-summary": "Boolean Output"
                    },
                    "datetimeOutput": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Datetime Output",
                      "x-ntx-summary": "Datetime Output"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/async-callback": {
      "post": {
        "summary": "POST async-callback",
        "description": "Calls back with different outcomes",
        "x-ntx-long-running-operation": true,
        "x-ntx-notification-content": {
          "description": "Result",
          "schema": {
            "$ref": "#/definitions/asyncCallbackResult"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "callbackUrl",
            "x-ntx-summary": "callback url",
            "description": "callback url",
            "x-ntx-notification-url": true,
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "requestBody",
            "description": "Request Body",
            "x-ntx-summary": "Request Body",
            "schema": {
              "type": "object",
              "properties": {
                "initialStatusCode": {
                  "type": "integer",
                  "description": "Initial status code to be returned",
                  "x-ntx-summary": "Initial Status Code"
                },
                "textInput": {
                  "type": "string",
                  "description": "Text Input",
                  "x-ntx-summary": "Text Input"
                },
                "resultStatus": {
                  "type": "string",
                  "description": "Status to be returned in result",
                  "x-ntx-summary": "Status to be returned in result"
                },
                "errorMessage": {
                  "type": "string",
                  "description": "Error message to be returned",
                  "x-ntx-summary": "Error message to be returned"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result after asynchronous processing",
            "schema": {
              "$ref": "#/definitions/asyncCallbackResult"
            }
          },
          "202": {
            "description": "Accepted for processing",
            "schema": {
              "type": "object",
              "properties": {
                "receiptId": {
                  "type": "string",
                  "description": "Receipt Id",
                  "x-ntx-summary": "Receipt Id"
                }
              }
            }
          }
        }
      }
    },
    "/sleep": {
      "get": {
        "summary": "GET response after 75 seconds",
        "description": "GET response after 75 seconds",
        "parameters": [
          {
            "type": "string",
            "name": "custom1",
            "in": "header",
            "description": "random unused header value"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/data/array/integer": {
      "get": {
        "summary": "Returns array of integers",
        "description": "Returns array of integers",
        "parameters": [
          {
            "in": "query",
            "name": "elements",
            "description": "Number of elements to be returned",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "echoRequest": {
      "type": "object"
    },
    "echoResponse": {
      "type": "object"
    },
    "rootBody": {
      "type": "object",
      "properties": {
        "allTypesInputs": {
          "type": "object",
          "properties": {
            "textInput": {
              "type": "string",
              "description": "Text Input",
              "x-ntx-summary": "Text Input"
            },
            "decimalInput": {
              "type": "number",
              "description": "Decimal Input",
              "x-ntx-summary": "Decimal Input"
            },
            "integerInput": {
              "type": "integer",
              "description": "Integer Input",
              "x-ntx-summary": "Integer Input"
            },
            "booleanInput": {
              "type": "boolean",
              "description": "Boolean Input",
              "x-ntx-summary": "Boolean Input"
            },
            "datetimeInput": {
              "type": "string",
              "format": "date-time",
              "description": "Datetime Input",
              "x-ntx-summary": "Datetime Input"
            },
            "collectionInput": {
              "type": "array",
              "items": {
                "type": "string",
                "description": "Collection Item Input"
              },
              "description": "Collection Input",
              "x-ntx-summary": "Collection Input"
            }
          }
        }
      }
    },
...
  }
}